WIP: Add a OVB te-broker server
Update tripleo-ci to run on an OVB cloud

This iteration of this patch uses the heat-templates in the
openstack-virtual-baremetal repository.

Some previous work I had been doing was to create heat templates
specific to toci, I believe thise a simpler as some can be remove
that we don't need. I've left the templates behind in heat-templates/ovb-*

Either way, I'm halting work here untill the question of weither me move
to tripleo-ci or not is answered as that will dictate how this takes shape.

Change-Id: Ie2d1c607f283e6babb00ea19d32bebae5383867a